refactor: Reorganize Zoom package structure to separate API concerns from business logic
- Move API-specific code to internal/infrastructure/zoom/api/ package
- Separate Zoom API client (pure API calls) from business logic provider
- Create ZoomProvider that wraps API client with domain-specific logic
- Maintain clean separation of concerns between API layer and domain logic
- Update imports and dependencies to reflect new structure
