--- internal/infrastructure/zoom/users.go
@@ -1,137 +1,0 @@
-// Copyright The Linux Foundation and each contributor to LFX.
-// SPDX-License-Identifier: MIT
-
-package zoom
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"io"
-	"log/slog"
-	"net/http"
-
-	"github.com/linuxfoundation/lfx-v2-meeting-service/internal/logging"
-)
-
-// ZoomUser represents a user in the Zoom account
-type ZoomUser struct {
-	ID        string `json:"id"`
-	Email     string `json:"email"`
-	FirstName string `json:"first_name"`
-	LastName  string `json:"last_name"`
-	Type      int    `json:"type"`
-	Status    string `json:"status"`
-}
-
-// ZoomUsersResponse represents the response from the users API
-type ZoomUsersResponse struct {
-	PageCount   int        `json:"page_count"`
-	PageNumber  int        `json:"page_number"`
-	PageSize    int        `json:"page_size"`
-	TotalRecord int        `json:"total_records"`
-	Users       []ZoomUser `json:"users"`
-}
-
-// User type constants for Zoom API
-const (
-	UserTypeBasic    = 1
-	UserTypeLicensed = 2
-	UserTypeOnPrem   = 3
-)
-
-// User status constants for Zoom API
-const (
-	UserStatusActive   = "active"
-	UserStatusInactive = "inactive"
-	UserStatusPending  = "pending"
-)
-
-// GetUsers retrieves users from the Zoom account
-func (c *Client) GetUsers(ctx context.Context) ([]ZoomUser, error) {
-	ctx = logging.AppendCtx(ctx, slog.String("zoom_operation", "get_users"))
-
-	resp, err := c.doRequest(ctx, http.MethodGet, "/users?status=active&page_size=100", nil)
-	if err != nil {
-		slog.ErrorContext(ctx, "failed to get Zoom users", logging.ErrKey, err)
-		return nil, err
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		body, _ := io.ReadAll(resp.Body)
-		err := parseErrorResponse(body)
-		slog.ErrorContext(ctx, "Zoom API returned error", logging.ErrKey, err, "status", resp.StatusCode)
-		return nil, err
-	}
-
-	var usersResp ZoomUsersResponse
-	if err := json.NewDecoder(resp.Body).Decode(&usersResp); err != nil {
-		slog.ErrorContext(ctx, "failed to decode users response", logging.ErrKey, err)
-		return nil, fmt.Errorf("failed to decode users response: %w", err)
-	}
-
-	slog.InfoContext(ctx, "successfully retrieved Zoom users",
-		"user_count", len(usersResp.Users),
-		"total_records", usersResp.TotalRecord)
-
-	return usersResp.Users, nil
-}
-
-// GetFirstAvailableUser gets the first active licensed user from the account
-func (c *Client) GetFirstAvailableUser(ctx context.Context) (*ZoomUser, error) {
-	ctx = logging.AppendCtx(ctx, slog.String("zoom_operation", "get_first_available_user"))
-
-	users, err := c.GetUsers(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get users: %w", err)
-	}
-
-	if len(users) == 0 {
-		return nil, fmt.Errorf("no users found in Zoom account")
-	}
-
-	// Find first active licensed user (type 2 = licensed)
-	for _, user := range users {
-		if user.Status == UserStatusActive && user.Type == UserTypeLicensed {
-			slog.InfoContext(ctx, "selected Zoom user for meeting creation",
-				"user_id", user.ID,
-				"email", user.Email,
-				"name", fmt.Sprintf("%s %s", user.FirstName, user.LastName))
-			return &user, nil
-		}
-	}
-
-	// If no licensed users, fall back to any active user
-	for _, user := range users {
-		if user.Status == UserStatusActive {
-			slog.InfoContext(ctx, "selected Zoom user for meeting creation (fallback)",
-				"user_id", user.ID,
-				"email", user.Email,
-				"name", fmt.Sprintf("%s %s", user.FirstName, user.LastName),
-				"type", user.Type)
-			return &user, nil
-		}
-	}
-
-	return nil, fmt.Errorf("no active users found in Zoom account")
-}
-
-// GetCachedUser gets the cached user or fetches the first available user if not cached
-func (c *Client) GetCachedUser(ctx context.Context) (*ZoomUser, error) {
-	if c.userCacheValid && c.cachedUser != nil {
-		slog.DebugContext(ctx, "using cached Zoom user", "user_id", c.cachedUser.ID)
-		return c.cachedUser, nil
-	}
-
-	user, err := c.GetFirstAvailableUser(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	// Cache the user for future use
-	c.cachedUser = user
-	c.userCacheValid = true
-
-	return user, nil
-}
